Slight Error
According to the comment "this runs for the no. of trials in the gdf file", it should be taking the breadth of X1, not length of X1. The length of X1 and ClassLnth(1) are the same, and so both the loops i and j were going to the same value. This didn't matter while implementing algo 1 and 2 because the number of trials matched length(X1) for some specific cases.